requirements.txt
paramiko
pyyaml
jinja2
